requirements.txt
fastapi
uvicorn[standard]

sqlmodel
psycopg2-binary

python-dotenv

bcrypt
python-jose

pydantic

# Whisper + audio processing
git+https://github.com/openai/whisper.git
torch
ffmpeg-python
numpy

python-multipart
# Groq LLM (scoring)
groq

# CORS
starlette
